requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile ./requirements.in
#
easyocr==1.7.2
    # via -r requirements.in
filelock==3.20.1
    # via torch
fsspec==2025.12.0
    # via torch
imageio==2.37.2
    # via scikit-image
jinja2==3.1.6
    # via torch
lazy-loader==0.4
    # via scikit-image
markupsafe==3.0.3
    # via jinja2
mpmath==1.3.0
    # via sympy
networkx==3.4.2
    # via
    #   scikit-image
    #   torch
ninja==1.13.0
    # via easyocr
numpy==2.2.6
    # via
    #   easyocr
    #   imageio
    #   opencv-python-headless
    #   scikit-image
    #   scipy
    #   shapely
    #   tifffile
    #   torchvision
nvidia-cublas-cu12==12.8.4.1
    # via
    #   nvidia-cudnn-cu12
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cuda-cupti-cu12==12.8.90
    # via torch
nvidia-cuda-nvrtc-cu12==12.8.93
    # via torch
nvidia-cuda-runtime-cu12==12.8.90
    # via torch
nvidia-cudnn-cu12==9.10.2.21
    # via torch
nvidia-cufft-cu12==11.3.3.83
    # via torch
nvidia-cufile-cu12==1.13.1.3
    # via torch
nvidia-curand-cu12==10.3.9.90
    # via torch
nvidia-cusolver-cu12==11.7.3.90
    # via torch
nvidia-cusparse-cu12==12.5.8.93
    # via
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cusparselt-cu12==0.7.1
    # via torch
nvidia-nccl-cu12==2.27.5
    # via torch
nvidia-nvjitlink-cu12==12.8.93
    # via
    #   nvidia-cufft-cu12
    #   nvidia-cusolver-cu12
    #   nvidia-cusparse-cu12
    #   torch
nvidia-nvshmem-cu12==3.3.20
    # via torch
nvidia-nvtx-cu12==12.8.90
    # via torch
opencv-python-headless==4.12.0.88
    # via easyocr
packaging==25.0
    # via
    #   lazy-loader
    #   scikit-image
pillow==12.0.0
    # via
    #   easyocr
    #   imageio
    #   scikit-image
    #   torchvision
pyclipper==1.4.0
    # via easyocr
python-bidi==0.6.7
    # via easyocr
pyyaml==6.0.3
    # via easyocr
scikit-image==0.25.2
    # via easyocr
scipy==1.15.3
    # via
    #   easyocr
    #   scikit-image
shapely==2.1.2
    # via easyocr
sympy==1.14.0
    # via torch
tifffile==2025.5.10
    # via scikit-image
torch==2.9.1
    # via
    #   easyocr
    #   torchvision
torchvision==0.24.1
    # via easyocr
tqdm==4.67.1
    # via -r requirements.in
triton==3.5.1
    # via torch
typing-extensions==4.15.0
    # via torch
